refactor: move validate_path to datasets
