+++ packages/ref/tests/unit/cli/test_ingest.py
@@ -1,45 +1,12 @@
 from pathlib import Path
 
-import pytest
-from ref_core.exceptions import OutOfTreeDatasetException
 from typer.testing import CliRunner
 
 from ref.cli import app
-from ref.cli.ingest import validate_path
 from ref.models import Dataset
 from ref.models.dataset import CMIP6Dataset, CMIP6File
 
 runner = CliRunner()
-
-
-def test_validate_prefix_with_valid_relative_path(config):
-    config.paths.data = Path("/data")
-
-    raw_path = "/data/subfolder/file.csv"
-    expected_path = Path("subfolder/file.csv")
-
-    result = validate_path(config, raw_path)
-    assert result == expected_path
-
-
-def test_validate_prefix_with_allow_out_of_tree_datasets(config):
-    config.paths.data = Path("/data")
-    config.paths.allow_out_of_tree_datasets = True
-
-    raw_path = "/other_dir/file.csv"
-    expected_path = Path("/other_dir/file.csv")
-
-    result = validate_path(config, raw_path)
-    assert result == expected_path
-
-
-def test_validate_prefix_with_invalid_relative_path(config):
-    config.paths.data = Path("/data")
-    config.paths.allow_out_of_tree_datasets = False
-
-    raw_path = "/other_dir/file.csv"
-    with pytest.raises(OutOfTreeDatasetException):
-        validate_path(config, raw_path)
 
 
 def test_ingest_help():
